% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RacTiter.R
\name{RacTiter}
\alias{RacTiter}
\title{\if{html}{\out{
<Add Title>
}}}
\usage{
RacTiter(
  titerdata,
  ag_order,
  ag_sequences = NULL,
  mode = "lineplot",
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
\if{html}{\out{
<Add Description>
}}
}
